// types/profile.ts
export type Profile = {
  profile: {
        id: string
        salonId: string
        name: string
        address: string
        phone: string
        description: string
        businessType: string
        portfolio: string
        priceRange: string
        openTime: string
        totalStaff: string
        tag: string[]
        createdAt: string
    }
    email: string
    isVerified: boolean
}